// fe/src/lib/api/generated/model/goodMoralStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type GoodMoralStatusValue = typeof GoodMoralStatusValue[keyof typeof GoodMoralStatusValue];


export const GoodMoralStatusValue = {
  REQUESTED: 'REQUESTED',
  ISSUED: 'ISSUED',
} as const;