// fe/src/lib/api/generated/model/goodMoralVariantValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type GoodMoralVariantValue = typeof GoodMoralVariantValue[keyof typeof GoodMoralVariantValue];


export const GoodMoralVariantValue = {
  CURRENT_STUDENT: 'CURRENT_STUDENT',
  GRADUATE: 'GRADUATE',
} as const;